+++ src/Core/IBus.cs
@@ -1,0 +1,8 @@
+// SPDX-FileCopyrightText: Copyright (c) 2025 Logan Bussell
+// SPDX-License-Identifier: MIT
+
+namespace NesNes.Core;
+
+public interface IBus : ICpuReadable, ICpuWritable
+{
+}
