+++ src/Core/NametableArrangement.cs
@@ -1,0 +1,10 @@
+// SPDX-FileCopyrightText: Copyright (c) 2025 Logan Bussell
+// SPDX-License-Identifier: MIT
+
+namespace NesNes.Core;
+
+public enum NametableArrangement
+{
+    Vertical,
+    Horizontal,
+}
