+++ src/Core/PpuConsts.cs
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: Copyright (c) 2025 Logan Bussell
+// SPDX-License-Identifier: MIT
+
+namespace NesNes.Core;
+
+internal static class PpuConsts
+{
+    public const int ChrRomSize = 0x2000; // 8KB
+}
